clippy

+++ pallets/rc-migrator/src/asset_rate.rs
@@ -50,7 +50,7 @@ impl<T: Config> PalletMigration for AssetRateMigrator<T> {
 				}
 			}
 			if T::MaxAhWeight::get()
-				.any_lt(T::AhWeightInfo::receive_asset_rates((messages.len() + 1) as u32))
+				.any_lt(T::AhWeightInfo::receive_asset_rates(messages.len() + 1))
 			{
 				log::info!(
 					target: LOG_TARGET,
