+++ pallets/rc-migrator/src/vesting.rs
@@ -76,7 +76,7 @@ impl<T: Config> PalletMigration for VestingMigrator<T> {
 				}
 			}
 			if T::MaxAhWeight::get()
-				.any_lt(T::AhWeightInfo::receive_vesting_schedules((messages.len() + 1) as u32))
+				.any_lt(T::AhWeightInfo::receive_vesting_schedules(messages.len() + 1))
 			{
 				log::info!(
 					target: LOG_TARGET,
